refactor(backend): moved files
+ moved files from backend/app to backend folder
+ fixed imports
+ fixed dev/start scripts (run as module now)
+ added root gitignore
+ added install.sh to prepare backend venv
+ added vscode basic settings for code-server
