+++ .env.example
@@ -7,7 +7,7 @@
 # in frontend's environment.*.ts, as well as nginx.conf
 API_PATH=/api
 
-# App log level, mainly for the unicorn command in start.py
+# App log level, mainly for the uvicorn command in start.py
 # critical, error, warning, info, debug, trace
 LOG_LEVEL=warning
