+++ backend/core/smtp.py
@@ -2,7 +2,7 @@ import smtplib
 from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
 from typing import Optional
-from app.config import Config
+from backend.config import Config
 import logging
