+++ backend/db/session.py
@@ -1,4 +1,4 @@
-from app.config import Config
+from backend.config import Config
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession, async_sessionmaker
 from typing import AsyncGenerator
 from sqlalchemy.ext.asyncio import AsyncSession
