+++ backend/schemas/user.py
@@ -1,7 +1,7 @@
 from typing import Optional, Self
 from pydantic import BaseModel, EmailStr, field_validator, model_validator, ConfigDict
 from .validators import password_validator
-from app.enums import EUserRole
+from backend.enums import EUserRole
 from datetime import datetime
