+++ install.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+python -m venv .venv
+source .venv/bin/activate
+pip install -r backend/requirements.txt
